Initial implementation of the log server
This implements the log server with Hapi.js.
It uses Glue to configure the server + plugins.
The api itself is implemented as a Hapi plugin.
It exposes a single endpoint /log where you may
POST JSON objects you want to log.

It uses Swagger (and Hapi's routes and Joi validators)
to generate documentation that can be viewed in a
swagger view.